+++ useintest/modules/irods/models.py
@@ -1,9 +1,9 @@
-from typing import List, Any
+from typing import Any
 
 import semantic_version
 
 from useintest.common import UseInTestModel
-from useintest.services.models import DockerisedService
+from useintest.services.models import DockerisedServiceWithUsers, User
 
 # Import from semantic version library
 Version = semantic_version.Version
@@ -11,7 +11,7 @@ Version = semantic_version.Version
 
 class IrodsResource(UseInTestModel):
     """
-    Model of a iRODS server resource.
+    An iRODS server resource.
     """
     def __init__(self, name: str, host: str, location: str):
         self.name = name
@@ -19,32 +19,28 @@ class IrodsResource(UseInTestModel):
         self.location = location
 
 
-# TODO: Extend User in /common
-class IrodsUser(UseInTestModel):
+class IrodsUser(User):
     """
-    Model of an iRODS user.
+    An iRODS user.
     """
     def __init__(self, username: str, zone: str, password: str=None, admin: bool=False):
-        super().__init__()
-        self.username = username
-        self.password = password
+        super().__init__(username, password)
         self.zone = zone
         self.admin = admin
 
     def __eq__(self, other: Any) -> bool:
-        return type(other) == type(self) \
-               and other.username == self.username \
-               and other.password == self.password \
+        return super().__eq__(other) \
                and other.zone == self.zone \
                and other.admin == self.admin
 
+    def __hash__(self) -> hash:
+        return hash(str(super().__hash__()) + self.zone + str(self.admin))
 
-# TODO: Extend ServiceWithUsers in /common
-class IrodsDockerisedService(DockerisedService):
+
+class IrodsDockerisedService(DockerisedServiceWithUsers):
     """
-    Model of a iRODS service running in Docker.
+    An iRODS service running in Docker.
     """
     def __init__(self):
         super().__init__()
-        self.users: List[IrodsUser] = []
         self.version: Version = None